feat: 9.9 xxxxxxxx

+++ xiaohashu-user/xiaohashu-user-biz/src/main/java/com/quanxiaoha/xiaohashu/user/biz/enums/SexEnum.java
@@ -15,7 +15,7 @@ public enum SexEnum {
 
     private final Integer value;
 
-    public static boolean isValid(Integer value) {
+    public static boolean isValid(Byte value) {
         for (SexEnum loginTypeEnum : SexEnum.values()) {
             if (Objects.equals(value, loginTypeEnum.getValue())) {
                 return true;
